+++ lib/design_system/widgets/buttons/secondary/defult_secondary_button.dart
@@ -1,15 +1,15 @@
 import 'package:flutter/material.dart';
-import 'package:moneyplus/design_system/component/buttons/money_button.dart';
+import 'package:moneyplus/design_system/widgets/buttons/money_button.dart';
 import 'package:moneyplus/design_system/theme/money_colors.dart';
 
-class DefaultErrorButton extends StatefulWidget {
+class DefaultSecondaryButton extends StatefulWidget {
   final String text;
   final VoidCallback? onPressed;
   final String? iconPath;
   final bool isLoading;
   final bool isEnabled;
 
-  const DefaultErrorButton({
+  const DefaultSecondaryButton({
     super.key,
     required this.text,
     this.onPressed,
@@ -19,10 +19,10 @@ class DefaultErrorButton extends StatefulWidget {
   });
 
   @override
-  State<DefaultErrorButton> createState() => _DefaultErrorButtonState();
+  State<DefaultSecondaryButton> createState() => _DefaultSecondaryButtonState();
 }
 
-class _DefaultErrorButtonState extends State<DefaultErrorButton> {
+class _DefaultSecondaryButtonState extends State<DefaultSecondaryButton> {
   @override
   Widget build(BuildContext context) {
     return MoneyButton(
@@ -31,9 +31,9 @@ class _DefaultErrorButtonState extends State<DefaultErrorButton> {
       iconPath: widget.iconPath,
       isLoading: widget.isLoading,
       isEnabled: widget.isEnabled,
-      backgroundColor: MoneyColors.light.redVariant,
+      backgroundColor: MoneyColors.light.surfaceLow,
       disabledBackgroundColor: MoneyColors.light.disabled,
-      textColor: MoneyColors.light.red,
+      textColor: MoneyColors.light.title,
       disabledTextColor: MoneyColors.light.onPrimary,
       hasShadow: false,
       fontSize: 14,
